removed unneeded cast

+++ src/tbprobe.c
@@ -1924,12 +1924,12 @@ JNIEXPORT jint JNICALL Java_com_winkelhagen_chess_syzygy_SyzygyBridge_getTBLarge
 
 JNIEXPORT jint JNICALL Java_com_winkelhagen_chess_syzygy_SyzygyBridge_probeWDL
   (JNIEnv * env, jclass clz, jlong white, jlong black, jlong kings, jlong queens, jlong rooks, jlong bishops, jlong knights, jlong pawns, jint ep, jboolean turn){
-    return tb_probe_wdl_impl(static_cast<uint64_t>(white), static_cast<uint64_t>(black), static_cast<uint64_t>(kings), static_cast<uint64_t>(queens), static_cast<uint64_t>(rooks), static_cast<uint64_t>(bishops), static_cast<uint64_t>(knights), static_cast<uint64_t>(pawns), ep, turn);
+    return tb_probe_wdl_impl(white, black, kings, queens, rooks, bishops, knights, pawns, ep, turn);
 }
 
 JNIEXPORT jint JNICALL Java_com_winkelhagen_chess_syzygy_SyzygyBridge_probeDTZ
   (JNIEnv * env, jclass clz, jlong white, jlong black, jlong kings, jlong queens, jlong rooks, jlong bishops, jlong knights, jlong pawns, jint rule50, jint ep, jboolean turn){
-    return tb_probe_root_impl(static_cast<uint64_t>(white), static_cast<uint64_t>(black), static_cast<uint64_t>(kings), static_cast<uint64_t>(queens), static_cast<uint64_t>(rooks), static_cast<uint64_t>(bishops), static_cast<uint64_t>(knights), static_cast<uint64_t>(pawns), rule50, ep, turn, NULL);
+    return tb_probe_root_impl(white, black, kings, queens, rooks, bishops, knights, pawns, rule50, ep, turn, NULL);
 }
